Fixed: conversion to np.float64 in list2interval().

+++ src/jpt/base/utils.py
@@ -158,8 +158,8 @@ def list2interval(l):
     Converts a list representation of an interval to an instance of type
     '''
     lower, upper = l
-    return ContinuousSet(np.NINF if lower in (np.NINF, -float('inf'), None) else lower,
-                         np.PINF if upper in (np.PINF, float('inf'), None) else upper)
+    return ContinuousSet(np.NINF if lower in (np.NINF, -float('inf'), None) else np.float64(lower),
+                         np.PINF if upper in (np.PINF, float('inf'), None) else np.float64(upper))
 
 
 def normalized(dist, identity_on_zeros=False, allow_neg=False):
